fix a possible overflow in showing CPU info on 32-bit systems

+++ base/sysinfo.jl
@@ -120,7 +120,7 @@ function _show_cpuinfo(io::IO, info::Sys.CPUinfo, header::Bool=true, prefix::Abs
     print(io, prefix)
     unit = tck > 0 ? " s  " : "    "
     tc = max(tck, 1)
-    d(i, unit=unit) = lpad(string(round(Int,i)), 9) * unit
+    d(i, unit=unit) = lpad(string(round(Int64,i)), 9) * unit
     print(io,
           lpad(string(info.speed), 5), " MHz  ",
           d(info.cpu_times!user / tc), d(info.cpu_times!nice / tc), d(info.cpu_times!sys / tc),
